+++ workspaces/app-defaults/plugins/app-react/src/alpha.ts
@@ -21,3 +21,4 @@
  */
 
 export * from './drawer';
+export * from './templateCard';